Introduces strided and offset tensors
This is a huge change that allows much more flexibility in the memory
layout. This will eventually lead to speed improvements, but this update
is so far making quite a few things a lot slower. This will have to be
improved over time.

Many things can be also have been broken that are not covered by tests,
so developing more tests is high priority.